+++ files/perks/mortality/data.lua
@@ -1,5 +1,6 @@
 local ns = dofile_once("mods/io__github__arcensoth__challenge_buffet/files/scripts/utils/namespacing.lua")
 local log = dofile_once(ns.file("scripts/utils/logging.lua"))
+local tags = dofile_once(ns.file("scripts/const/tags.lua"))
 
 local data = {}
 
@@ -21,7 +22,13 @@ data.update_hp = function(entity, damage)
 		for i, damagemodel in ipairs(damagemodels) do
 			local current_hp = tonumber(ComponentGetValue(damagemodel, "hp"))
 			local new_hp = current_hp - damage
+			log.debug("Setting max HP to match new HP " .. new_hp .. " after taking " .. damage .. " damage")
 			ComponentSetValue(damagemodel, "max_hp", new_hp)
+			-- Make sure to check whether max HP cap should be synced with max HP.
+			if (EntityHasTag(entity, tags.sync_max_hp_with_cap)) then
+				log.debug("Also syncing max HP cap with new max HP: " .. new_hp)
+				ComponentSetValue(damagemodel, "max_hp_cap", new_hp)
+			end
 		end
 	end
 end