+++ files/scripts/const/tags.lua
@@ -1,0 +1,5 @@
+local ns = dofile_once("mods/io__github__arcensoth__challenge_buffet/files/scripts/utils/namespacing.lua")
+
+return {
+    sync_max_hp_with_cap = ns.key("sync_max_hp_with_cap"),
+}