+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed Mortality not synchronizing max HP cap to max HP when Heartbreak is also present
+
 ## [0.1.0] - 2020-09-12
 ### Added
 - Added 3 unique challenges in the form of "cursed" perks